Add auth GRPC

+++ auth/pkg/models/session.go
@@ -1,6 +1,6 @@
 package models
 
 type Session struct {
-	UserId int    `json:"userId"`
+	UserId int64  `json:"userId"`
 	Token  string `json:"-"`
 }
